feat: file card layout tweaks
