+++ src/sections/collection/collection-items-panel/items-list/file-card/FileCard.module.scss
@@ -87,22 +87,16 @@
   }
 }
 
-.card-info-container {
+.card-body-container {
   align-self: flex-start;
   font-size: $dv-font-size-sm;
 
-  .date-link-wrapper {
-    display: flex;
-    flex-wrap: wrap;
-    column-gap: 0.5rem;
+  .release-or-create-date {
+    color: $dv-subtext-color;
+  }
 
-    .date {
-      color: $dv-subtext-color;
-    }
-
-    a {
-      @include link-hover-underlined;
-    }
+  a {
+    @include link-hover-underlined;
   }
 
   .info {
@@ -123,8 +117,9 @@
     line-clamp: 3;
     -webkit-box-orient: vertical;
     width: 100%;
+    margin: 0;
     overflow: hidden;
-    color: black;
+    color: #000;
   }
 }
