+++ src/sections/collection/collection-items-panel/items-list/file-card/FileCard.tsx
@@ -1,7 +1,6 @@
 import { FileItemTypePreview } from '@/files/domain/models/FileItemTypePreview'
 import { FileCardHeader } from './FileCardHeader'
-import { FileCardThumbnail } from './FileCardThumbnail'
-import { FileCardInfo } from './FileCardInfo'
+import { FileCardBody } from './FileCardBody'
 import styles from './FileCard.module.scss'
 
 interface FileCardProps {
@@ -12,10 +11,7 @@ export function FileCard({ filePreview }: FileCardProps) {
   return (
     <article className={styles['card-main-container']} data-testid="file-card">
       <FileCardHeader filePreview={filePreview} />
-      <div className={styles['thumbnail-and-info-wrapper']}>
-        <FileCardThumbnail filePreview={filePreview} />
-        <FileCardInfo filePreview={filePreview} />
-      </div>
+      <FileCardBody filePreview={filePreview} />
     </article>
   )
 }
